+++ src/components/movie/MovieItem.tsx
