+++ src/components/movie/MoviesSlider.tsx
