+++ src/components/movie/PosterImage.tsx
