+++ src/components/movie/PosterLabel.tsx
