+++ src/components/movie/list/types.ts
@@ -1,0 +1,7 @@
+export enum MovieListMode {
+  section,
+  search,
+  discover,
+}
+
+export type MovieListModeKey = keyof typeof MovieListMode;
