+++ src/components/pages/movie/detail/index.tsx
@@ -1,141 +1,29 @@
 import {
   AspectRatio,
-  Avatar,
   Badge,
   Button,
   Flex,
-  FormControl,
   Grid,
   Heading,
-  Input,
   Link as ChakraLink,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalFooter,
-  ModalHeader,
-  ModalOverlay,
   Skeleton,
   Text,
   useColorMode,
-  useDisclosure,
 } from "@chakra-ui/react";
 import Head from "next/head";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import { ChangeEvent, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { BiLinkExternal } from "react-icons/bi";
 import { FaImdb } from "react-icons/fa";
 import { GrGallery } from "react-icons/gr";
 
 import Error from "components/layout/Error";
-import PosterImage, { IMAGE_URL } from "components/movies/PosterImage";
+import CastsWrapper from "components/movie/detail/CastsWrapper";
+import PosterImage from "components/movie/PosterImage";
 import { useMovieCredits } from "services/tmdb/movie/credits";
-import { MovieCreditsResponse } from "services/tmdb/movie/credits/types";
 import { useMovieDetail } from "services/tmdb/movie/detail";
 import { convertToPrice } from "utils/convertToPrice";
-
-type CastsWrapperProps = {
-  isLoadingCredits: boolean;
-  credits?: MovieCreditsResponse;
-};
-
-const CastsWrapper = ({ isLoadingCredits, credits }: CastsWrapperProps) => {
-  const { isOpen, onOpen, onClose } = useDisclosure();
-
-  const [keyword, setKeyword] = useState<string>("");
-
-  const handleChangeKeyword = (event: ChangeEvent<HTMLInputElement>) =>
-    setKeyword(event.target.value);
-
-  return (
-    <Skeleton isLoaded={!isLoadingCredits}>
-      {credits && (
-        <Flex alignItems="center" gridGap={3} minHeight={24} overflowX="scroll">
-          <Button padding={8} borderRadius="50%" onClick={onOpen}>
-            all
-          </Button>
-          {credits.cast.slice(0, 20).map((movieCast) => (
-            <Link
-              href={`/person/${movieCast.id}`}
-              key={`${movieCast.name}-${movieCast.id}`}
-              passHref
-            >
-              <Avatar
-                cursor="pointer"
-                size="lg"
-                src={`${IMAGE_URL}${movieCast.profile_path}`}
-                name={movieCast.name}
-              />
-            </Link>
-          ))}
-          <Button padding={8} borderRadius="50%" onClick={onOpen}>
-            more
-          </Button>
-
-          <Modal
-            isCentered
-            isOpen={isOpen}
-            onClose={onClose}
-            scrollBehavior="inside"
-          >
-            <ModalOverlay />
-
-            <ModalContent>
-              <ModalHeader>
-                <Heading>Casts</Heading>
-                <FormControl marginY={2}>
-                  <Input
-                    type="text"
-                    placeholder="search"
-                    value={keyword}
-                    onChange={handleChangeKeyword}
-                  />
-                </FormControl>
-              </ModalHeader>
-              <ModalCloseButton />
-
-              <ModalBody>
-                <Grid gap={4} templateColumns={["repeat(1, 1fr)"]}>
-                  {credits.cast
-                    .filter(
-                      (unfilteredCast) =>
-                        unfilteredCast.name
-                          .toLowerCase()
-                          .indexOf(keyword.toLowerCase()) > -1
-                    )
-                    .map((movieCast) => (
-                      <Link
-                        href={`/person/${movieCast.id}`}
-                        key={`${movieCast.name}-${movieCast.id}`}
-                        passHref
-                      >
-                        <Flex
-                          cursor="pointer"
-                          alignItems="center"
-                          gridColumnGap={2}
-                        >
-                          <Avatar
-                            size="lg"
-                            name={movieCast.name}
-                            src={`${IMAGE_URL}${movieCast.profile_path}`}
-                          />
-                          <Text>{movieCast.name}</Text>
-                        </Flex>
-                      </Link>
-                    ))}
-                </Grid>
-              </ModalBody>
-
-              <ModalFooter />
-            </ModalContent>
-          </Modal>
-        </Flex>
-      )}
-    </Skeleton>
-  );
-};
 
 // eslint-disable-next-line sonarjs/cognitive-complexity, complexity
 const MovieDetailPage = () => {
