+++ src/pages/movies/[section].tsx
@@ -1,4 +1,4 @@
-import MovieListContainer from "components/movies/MovieListContainer";
+import MovieListContainer from "components/movie/list";
 
 const MoviesBySection = () => {
   return <MovieListContainer listMode="section" />;
