+++ src/pages/movies/genre/[genre].tsx
@@ -1,4 +1,4 @@
-import MovieListContainer from "components/movies/MovieListContainer";
+import MovieListContainer from "components/movie/list";
 
 const Genre = () => {
   return <MovieListContainer listMode="discover" />;
