+++ src/pages/movies/search.tsx
@@ -1,4 +1,4 @@
-import MovieListContainer from "components/movies/MovieListContainer";
+import MovieListContainer from "components/movie/list";
 
 const MoviesSearch = () => {
   return <MovieListContainer listMode="search" />;
